rbac-server/requirements.txt
flask
mysql-connector-python
bcrypt
pytest